refactor(theme): centralize color foundations and palette structure
