+++ src/stories/ColorPalette.mdx
@@ -10,6 +10,7 @@ import { theme } from '@theme';
 <ColorPalette>
   <ColorItem title="Neutral" colors={{ 
       black: theme.palette.neutral.black,
+      white: theme.palette.neutral.white,
       850: theme.palette.neutral[850], 
       700: theme.palette.neutral[700], 
       650: theme.palette.neutral[650], 
@@ -18,10 +19,9 @@ import { theme } from '@theme';
       200: theme.palette.neutral[200], 
       100: theme.palette.neutral[100],
       50: theme.palette.neutral[50],
-      white: theme.palette.neutral.white
-    }} 
-  />
-  
+    }}
+/>
+
 <ColorItem
   title="Primary"
   colors={{
