+++ src/theme/foundations/colors.ts
@@ -54,6 +54,7 @@ const success = {
   850: '#224021',
   700: '#427940',
   500: '#6CC66A',
+  400: '#89D188',
   100: '#E1F4E1',
 };
